CODE NO:-1 (Find Armstrong Number)
This is the python code for finding any number whether it is an Armstrong number or not.


An Armstrong number is a number that is equal to the sum of its own digits, each raised to the power of the number of digits. For example, 153 is an Armstrong number because 1^3 + 5^3 + 3^3 = 153.